requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
discord.py==1.6.0
dnspython==2.1.0
idna==3.1
idna-ssl==1.1.0
multidict==5.1.0
Pillow==8.1.0
pymongo==3.11.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wincertstore==0.2
yarl==1.6.3
